pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "asyncio>=4.0.0",
    "flask>=3.1.3",
    "highrise-bot-sdk>=24.1.0",
]
